+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = '2022, Peter Sun'
 author = 'Peter Sun'
 
 # The full version, including alpha/beta/rc tags
-release = '0.3.1'
+release = '0.4.0'
 
 
 # -- General configuration ---------------------------------------------------